requirements.txt
Werkzeug==1.0.1
mtcnn==0.1.0
Flask_Cors==3.0.9
Keras==2.4.3
Flask_WTF==0.14.3
gdown==3.12.2
numpy==1.19.3
tensorflow
Flask==1.1.2
matplotlib==3.4.1
Pillow==8.2.0
scikit_learn==0.24.1
gunicorn
opencv-python-headless


